Add a dataset

Add a dataset to a study object

add_data(study, id, data = NULL, vardesc = list(), design = NULL, ...)

Arguments

study

A study list object with class scivrs_study

id

The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it

data

The dataset as a data.frame, codebook object, path to a data file, or path to a codebook file

vardesc

Optional variable properties (see codebook)

design

A faux design specification

...

Further dataset properties (see codebook)

Value

A study object with class scivrs_study

Examples

s <- study() %>% add_data("dat", iris)
#> Sepal.Length set to dataType float
#> Sepal.Width set to dataType float
#> Petal.Length set to dataType float
#> Petal.Width set to dataType float
#> Species set to dataType string
study_to_json(s)
#> {
#>     "name": "Demo Study",
#>     "info": [],
#>     "authors": [],
#>     "hypotheses": [],
#>     "methods": [],
#>     "data": [
#>         {
#>             "id": "dat",
#>             "codebook": {
#>                 "@context": "https://schema.org/",
#>                 "@type": "Dataset",
#>                 "name": "dat",
#>                 "schemaVersion": "Psych-DS 0.1.0",
#>                 "variableMeasured": [
#>                     {
#>                         "@type": "PropertyValue",
#>                         "name": "Sepal.Length",
#>                         "description": "Sepal.Length",
#>                         "dataType": "float"
#>                     },
#>                     {
#>                         "@type": "PropertyValue",
#>                         "name": "Sepal.Width",
#>                         "description": "Sepal.Width",
#>                         "dataType": "float"
#>                     },
#>                     {
#>                         "@type": "PropertyValue",
#>                         "name": "Petal.Length",
#>                         "description": "Petal.Length",
#>                         "dataType": "float"
#>                     },
#>                     {
#>                         "@type": "PropertyValue",
#>                         "name": "Petal.Width",
#>                         "description": "Petal.Width",
#>                         "dataType": "float"
#>                     },
#>                     {
#>                         "@type": "PropertyValue",
#>                         "name": "Species",
#>                         "description": "Species",
#>                         "dataType": "string",
#>                         "levels": {
#>                             "setosa": "setosa",
#>                             "versicolor": "versicolor",
#>                             "virginica": "virginica"
#>                         },
#>                         "levelsOrdered": false
#>                     }
#>                 ]
#>             },
#>             "data": {
#>                 "Sepal.Length": [5.1, 4.9, 4.7, 4.6, 5, 5.4, 4.6, 5, 4.4, 4.9, 5.4, 4.8, 4.8, 4.3, 5.8, 5.7, 5.4, 5.1, 5.7, 5.1, 5.4, 5.1, 4.6, 5.1, 4.8, 5, 5, 5.2, 5.2, 4.7, 4.8, 5.4, 5.2, 5.5, 4.9, 5, 5.5, 4.9, 4.4, 5.1, 5, 4.5, 4.4, 5, 5.1, 4.8, 5.1, 4.6, 5.3, 5, 7, 6.4, 6.9, 5.5, 6.5, 5.7, 6.3, 4.9, 6.6, 5.2, 5, 5.9, 6, 6.1, 5.6, 6.7, 5.6, 5.8, 6.2, 5.6, 5.9, 6.1, 6.3, 6.1, 6.4, 6.6, 6.8, 6.7, 6, 5.7, 5.5, 5.5, 5.8, 6, 5.4, 6, 6.7, 6.3, 5.6, 5.5, 5.5, 6.1, 5.8, 5, 5.6, 5.7, 5.7, 6.2, 5.1, 5.7, 6.3, 5.8, 7.1, 6.3, 6.5, 7.6, 4.9, 7.3, 6.7, 7.2, 6.5, 6.4, 6.8, 5.7, 5.8, 6.4, 6.5, 7.7, 7.7, 6, 6.9, 5.6, 7.7, 6.3, 6.7, 7.2, 6.2, 6.1, 6.4, 7.2, 7.4, 7.9, 6.4, 6.3, 6.1, 7.7, 6.3, 6.4, 6, 6.9, 6.7, 6.9, 5.8, 6.8, 6.7, 6.7, 6.3, 6.5, 6.2, 5.9],
#>                 "Sepal.Width": [3.5, 3, 3.2, 3.1, 3.6, 3.9, 3.4, 3.4, 2.9, 3.1, 3.7, 3.4, 3, 3, 4, 4.4, 3.9, 3.5, 3.8, 3.8, 3.4, 3.7, 3.6, 3.3, 3.4, 3, 3.4, 3.5, 3.4, 3.2, 3.1, 3.4, 4.1, 4.2, 3.1, 3.2, 3.5, 3.6, 3, 3.4, 3.5, 2.3, 3.2, 3.5, 3.8, 3, 3.8, 3.2, 3.7, 3.3, 3.2, 3.2, 3.1, 2.3, 2.8, 2.8, 3.3, 2.4, 2.9, 2.7, 2, 3, 2.2, 2.9, 2.9, 3.1, 3, 2.7, 2.2, 2.5, 3.2, 2.8, 2.5, 2.8, 2.9, 3, 2.8, 3, 2.9, 2.6, 2.4, 2.4, 2.7, 2.7, 3, 3.4, 3.1, 2.3, 3, 2.5, 2.6, 3, 2.6, 2.3, 2.7, 3, 2.9, 2.9, 2.5, 2.8, 3.3, 2.7, 3, 2.9, 3, 3, 2.5, 2.9, 2.5, 3.6, 3.2, 2.7, 3, 2.5, 2.8, 3.2, 3, 3.8, 2.6, 2.2, 3.2, 2.8, 2.8, 2.7, 3.3, 3.2, 2.8, 3, 2.8, 3, 2.8, 3.8, 2.8, 2.8, 2.6, 3, 3.4, 3.1, 3, 3.1, 3.1, 3.1, 2.7, 3.2, 3.3, 3, 2.5, 3, 3.4, 3],
#>                 "Petal.Length": [1.4, 1.4, 1.3, 1.5, 1.4, 1.7, 1.4, 1.5, 1.4, 1.5, 1.5, 1.6, 1.4, 1.1, 1.2, 1.5, 1.3, 1.4, 1.7, 1.5, 1.7, 1.5, 1, 1.7, 1.9, 1.6, 1.6, 1.5, 1.4, 1.6, 1.6, 1.5, 1.5, 1.4, 1.5, 1.2, 1.3, 1.4, 1.3, 1.5, 1.3, 1.3, 1.3, 1.6, 1.9, 1.4, 1.6, 1.4, 1.5, 1.4, 4.7, 4.5, 4.9, 4, 4.6, 4.5, 4.7, 3.3, 4.6, 3.9, 3.5, 4.2, 4, 4.7, 3.6, 4.4, 4.5, 4.1, 4.5, 3.9, 4.8, 4, 4.9, 4.7, 4.3, 4.4, 4.8, 5, 4.5, 3.5, 3.8, 3.7, 3.9, 5.1, 4.5, 4.5, 4.7, 4.4, 4.1, 4, 4.4, 4.6, 4, 3.3, 4.2, 4.2, 4.2, 4.3, 3, 4.1, 6, 5.1, 5.9, 5.6, 5.8, 6.6, 4.5, 6.3, 5.8, 6.1, 5.1, 5.3, 5.5, 5, 5.1, 5.3, 5.5, 6.7, 6.9, 5, 5.7, 4.9, 6.7, 4.9, 5.7, 6, 4.8, 4.9, 5.6, 5.8, 6.1, 6.4, 5.6, 5.1, 5.6, 6.1, 5.6, 5.5, 4.8, 5.4, 5.6, 5.1, 5.1, 5.9, 5.7, 5.2, 5, 5.2, 5.4, 5.1],
#>                 "Petal.Width": [0.2, 0.2, 0.2, 0.2, 0.2, 0.4, 0.3, 0.2, 0.2, 0.1, 0.2, 0.2, 0.1, 0.1, 0.2, 0.4, 0.4, 0.3, 0.3, 0.3, 0.2, 0.4, 0.2, 0.5, 0.2, 0.2, 0.4, 0.2, 0.2, 0.2, 0.2, 0.4, 0.1, 0.2, 0.2, 0.2, 0.2, 0.1, 0.2, 0.2, 0.3, 0.3, 0.2, 0.6, 0.4, 0.3, 0.2, 0.2, 0.2, 0.2, 1.4, 1.5, 1.5, 1.3, 1.5, 1.3, 1.6, 1, 1.3, 1.4, 1, 1.5, 1, 1.4, 1.3, 1.4, 1.5, 1, 1.5, 1.1, 1.8, 1.3, 1.5, 1.2, 1.3, 1.4, 1.4, 1.7, 1.5, 1, 1.1, 1, 1.2, 1.6, 1.5, 1.6, 1.5, 1.3, 1.3, 1.3, 1.2, 1.4, 1.2, 1, 1.3, 1.2, 1.3, 1.3, 1.1, 1.3, 2.5, 1.9, 2.1, 1.8, 2.2, 2.1, 1.7, 1.8, 1.8, 2.5, 2, 1.9, 2.1, 2, 2.4, 2.3, 1.8, 2.2, 2.3, 1.5, 2.3, 2, 2, 1.8, 2.1, 1.8, 1.8, 1.8, 2.1, 1.6, 1.9, 2, 2.2, 1.5, 1.4, 2.3, 2.4, 1.8, 1.8, 2.1, 2.4, 2.3, 1.9, 2.3, 2.5, 2.3, 1.9, 2, 2.3, 1.8],
#>                 "Species": [
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "setosa",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "versicolor",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica",
#>                     "virginica"
#>                 ]
#>             }
#>         }
#>     ],
#>     "analyses": []
#> }
#>