Run analysis

Run the analyses on the data

study_analyse(study)

study_analyze(study)

Arguments

study
A study list object with class reg_study

Value

A study object with class reg_study

Examples

s <- study() %>%
  add_hypothesis("H1", "Petal width and length will be positively correlated.") %>%
  add_analysis("A1", cor.test(dat$Petal.Width, dat$Petal.Length)) %>%
  add_criterion("sig", "p.value", "<", 0.05) %>%
  add_criterion("pos", "estimate", ">", 0) %>%
  add_eval("corroboration",
           "Petal width is significantly and positively correlated to length",
           "sig & pos") %>%
  add_eval("falsification",
           "Petal width is significantly and negatively correlated to length",
           "sig & !pos") %>%
  add_data("dat", iris) %>%
  study_analyse()

#> Hypothesis H1, Criterion sig: p.value < 0.05 is TRUE (p.value = 0)

#> Hypothesis H1, Criterion pos: estimate > 0 is TRUE (estimate = 0.96)

#> Hypothesis H1, Evaluation: corroborate

study_json(s)

#> {
#>     "name": "Demo Study",
#>     "info": [],
#>     "authors": [],
#>     "hypotheses": [
#>         {
#>             "id": "H1",
#>             "description": "Petal width and length will be positively correlated.",
#>             "criteria": [
#>                 {
#>                     "id": "sig",
#>                     "analysis_id": "A1",
#>                     "result": "p.value",
#>                     "operator": "<",
#>                     "comparator": 0.05,
#>                     "conclusion": true
#>                 },
#>                 {
#>                     "id": "pos",
#>                     "analysis_id": "A1",
#>                     "result": "estimate",
#>                     "operator": ">",
#>                     "comparator": 0,
#>                     "conclusion": true
#>                 }
#>             ],
#>             "corroboration": {
#>                 "description": "Petal width is significantly and positively correlated to length",
#>                 "evaluation": "sig & pos"
#>             },
#>             "falsification": {
#>                 "description": "Petal width is significantly and negatively correlated to length",
#>                 "evaluation": "sig & !pos"
#>             },
#>             "conclusion": "corroborate"
#>         }
#>     ],
#>     "methods": [],
#>     "data": [
#>         {
#>             "id": "dat",
#>             "@type": "Dataset",
#>             "schemaVersion": "Psych-DS 0.1.0",
#>             "variableMeasured": [
#>                 {
#>                     "type": "PropertyValue",
#>                     "unitText": "Sepal.Length",
#>                     "name": "Sepal.Length",
#>                     "missingValues": 0,
#>                     "minValue": 4.3,
#>                     "maxValue": 7.9,
#>                     "meanValue": 5.8433,
#>                     "sdValue": 0.8281
#>                 },
#>                 {
#>                     "type": "PropertyValue",
#>                     "unitText": "Sepal.Width",
#>                     "name": "Sepal.Width",
#>                     "missingValues": 0,
#>                     "minValue": 2,
#>                     "maxValue": 4.4,
#>                     "meanValue": 3.0573,
#>                     "sdValue": 0.4359
#>                 },
#>                 {
#>                     "type": "PropertyValue",
#>                     "unitText": "Petal.Length",
#>                     "name": "Petal.Length",
#>                     "missingValues": 0,
#>                     "minValue": 1,
#>                     "maxValue": 6.9,
#>                     "meanValue": 3.758,
#>                     "sdValue": 1.7653
#>                 },
#>                 {
#>                     "type": "PropertyValue",
#>                     "unitText": "Petal.Width",
#>                     "name": "Petal.Width",
#>                     "missingValues": 0,
#>                     "minValue": 0.1,
#>                     "maxValue": 2.5,
#>                     "meanValue": 1.1993,
#>                     "sdValue": 0.7622
#>                 },
#>                 {
#>                     "type": "PropertyValue",
#>                     "unitText": "Species",
#>                     "name": "Species",
#>                     "missingValues": 0,
#>                     "levels": [
#>                         "setosa",
#>                         "versicolor",
#>                         "virginica"
#>                     ]
#>                 }
#>             ]
#>         }
#>     ],
#>     "analyses": [
#>         {
#>             "id": "A1",
#>             "func": "analysis_A1_func",
#>             "code": [
#>                 "function () ",
#>                 "{",
#>                 "    cor.test(dat$Petal.Width, dat$Petal.Length)",
#>                 "}"
#>             ],
#>             "results": {
#>                 "statistic": 43.3872,
#>                 "parameter": 148,
#>                 "p.value": 4.675e-86,
#>                 "estimate": 0.9629,
#>                 "null.value": 0,
#>                 "alternative": "two.sided",
#>                 "method": "Pearson's product-moment correlation",
#>                 "data.name": "dat$Petal.Width and dat$Petal.Length",
#>                 "conf.int": [0.9491, 0.973]
#>             }
#>         }
#>     ]
#> }
#>