Add simulated data

Uses faux::sim_design() to generate a data table with a specified within and between design.

add_sim_data(
  study,
  data_id,
  within = list(),
  between = list(),
  n = 100,
  mu = 0,
  sd = 1,
  r = 0,
  empirical = FALSE,
  long = FALSE,
  dv = list(y = "value"),
  id = list(id = "id"),
  plot = FALSE,
  seed = NULL,
  rep = 1
)

Arguments

study
A study list object with class reg_study

data_id
The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it

within
a list of the within-subject factors

between
a list of the between-subject factors

n
the number of samples required

mu
the means of the variables

sd
the standard deviations of the variables

r
the correlations among the variables (can be a single number, full correlation matrix as a matric or vector, or a vector of the upper right triangle of the correlation matrix

empirical
logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance

long
Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format

dv
the name of the dv for long plots (defaults to y)

id
the name of the id column (defaults to id)

plot
whether to show a plot of the design

seed
a single value, interpreted as an integer, or NULL (see set.seed)

rep
the number of data frames to return (default 1); if greater than 1, the returned data frame is nested by rep

Value

A study object with class reg_study

Examples

s <- study() %>%
  add_sim_data("dat", within = list(time = c("day", "night")),
               between = list(pet = c("cat", "dog")),
               n = 10, mu = 100, sd = 10, r = 0.5,
               seed = 8675309)
s$data[[1]]$data

#>     id pet       day     night
#> 1  S01 cat  99.11711  83.62158
#> 2  S02 cat 101.13799 111.36437
#> 3  S03 cat  93.90440  95.40523
#> 4  S04 cat 120.87487 114.27523
#> 5  S05 cat 114.19972 104.25383
#> 6  S06 cat  98.68728 118.41187
#> 7  S07 cat 102.44677  98.02875
#> 8  S08 cat 110.33043 101.32406
#> 9  S09 cat 105.70718 104.20130
#> 10 S10 cat 111.96555 103.68661
#> 11 S11 dog 115.31684 119.07868
#> 12 S12 dog  94.61344 106.14874
#> 13 S13 dog  88.62430 104.37333
#> 14 S14 dog  92.96108  98.84635
#> 15 S15 dog  99.48275  93.33230
#> 16 S16 dog 107.06154 104.77243
#> 17 S17 dog 105.65773 106.29690
#> 18 S18 dog 106.18754 103.05281
#> 19 S19 dog  99.63078  97.14579
#> 20 S20 dog 104.08538 102.54743